Move chat prompts to prompts
